Better handling of temp directories.
pinto-server was throwing errors (from within Path::Class) about
not being able to chdir into a temp directory that it was trying
to remove.  After some trial and error, I deduced it was the temp
directory that Dist::Requires creates that caused the problem.

I still don't know exactly why the error occurred, but if I forcibly
remove the temp directory after extracting the prereqs, then the
problem goes a way.  This is actually better than letting the
directory sit around until the process exits anyway.